requirements.txt
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
black==22.3.0
click==8.1.2
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
importlib-metadata==4.11.3
iniconfig==1.1.1
Markdown==3.3.6
mock==4.0.3
model-bakery==1.5.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.8
pytest==7.1.2
pytest-django==4.5.2
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.2.0
zipp==3.8.0
